refactor(timepicker): decouple angular
